requirements.txt
setuptools~=68.0.0
pandas~=2.0.3
numpy~=1.25.1
tqdm~=4.65.0
Pillow~=10.0.0


pyfoliage~=0.0.2